requirements.txt
numpy>=1.24,<3
Pillow>=10,<12
torch>=2.2,<3
gymnasium[atari,accept-rom-license]>=1.0,<2
ale-py>=0.10,<1
